+++ strider/__data__.py
@@ -1,4 +1,4 @@
-__version__ = '0.6.4a.dev'
+__version__ = '0.6.4b'
 __author__ = 'ben avrahami'
 __copyright__ = 'Copyright (c) 2018 Ben Avrahami'
 __homepage__ = 'https://github.com/bentheiii/strider'