+++ strider/__init__.py
@@ -7,12 +7,5 @@ from strider.line_edit import LineEdit
 
 __dev__ = "dev" in __version__
 
-# todo quick-tags in command line
-# todo make tag order deterministic
-# todo does q, shift+q work without tags?
-# todo click in "h"
-# todo auto-walk
-# todo step-forward
-# todo license, readme
-# todo add changelog
 # todo bug
+# todo allow moving window/ zooming during auto-play